+++ designsafe/apps/api/systems/views.py
@@ -7,7 +7,7 @@ import logging
 import json
 from django.http import JsonResponse
 from designsafe.apps.api.views import AuthenticatedApiView
-from designsafe.utils.system_access import create_system_credentials
+from designsafe.apps.onboarding.steps.system_access_v3 import create_system_credentials
 from designsafe.utils.encryption import createKeyPair
 from .utils import add_pub_key_to_resource
