+++ designsafe/apps/onboarding/steps/system_access_v3.py
@@ -14,8 +14,13 @@ from tapipy.errors import BaseTapyException
 logger = logging.getLogger(__name__)
 
 
-def create_system_credentials(
-    client, username, public_key, private_key, system_id, skipCredentialCheck=False
+def create_system_credentials(  # pylint: disable=too-many-arguments
+    client,
+    username,
+    public_key,
+    private_key,
+    system_id,
+    skipCredentialCheck=False,  # pylint: disable=invalid-name
 ) -> int:
     """
     Set an RSA key pair as the user's auth credential on a Tapis system.
@@ -30,14 +35,16 @@ def create_system_credentials(
     )
 
 
-def register_public_key(username, publicKey, system_id) -> int:
+def register_public_key(
+    username, publicKey, system_id  # pylint: disable=invalid-name
+) -> int:
     """
     Push a public key to the Key Service API.
     """
     url = "https://api.tacc.utexas.edu/keys/v2/" + username
-    headers = {"Authorization": "Bearer {}".format(settings.KEY_SERVICE_TOKEN)}
+    headers = {"Authorization": f"Bearer {settings.KEY_SERVICE_TOKEN}"}
     data = {"key_value": publicKey, "tags": [{"name": "system", "value": system_id}]}
-    response = requests.post(url, json=data, headers=headers)
+    response = requests.post(url, json=data, headers=headers, timeout=60)
     response.raise_for_status()
     return response.status_code
